use the same getTestGateway function for all tests
